Refactor Home component from .jsx to .tsx
The Home.jsx component was refactored to Home.tsx. This change was made to improve the coding standard by switching to TypeScript. The Home.tsx component now uses media query to dynamically render Desktop or Categories based on the user's screen size. Furthermore, main.jsx was also adjusted to render Home component as the root path.