Add function to highlight substring in a string
A new utility function, `colorWordInString()`, has been added in utils.jsx which highlights a particular word in a string. This function enhances the visibility of particular keyword(s) in a string by wrapping them with a span element and a "highlight" class. This could be useful when drawing attention to certain words in a text output.