Adjust placement and padding of elements in CSS
Adjusted the top property of a fixed position element from 76% to 71% and reduced the padding-top value of 'mapInfoPane--container_venue p' element from 20px to 10px in the CSS style sheet. These changes were made to improve the positioning and alignment of the elements on the website for better user interface.