Adjust header and navigation CSS styles
The CSS styles have been adjusted to alter the positioning of the navigation header in `Desktop.tsx`. Moreover, `index.css` has been modified to fix the declared position of `.desktop-UI_nav` and to tweak the top and right properties. The changes enhance the visual appearance and placement of navigation items.