Refactor scrolling behavior and delay in Intersection Observer
Improved navigation experience by refactoring the scrolling behavior in the 'Categories' and 'TopList' components to scroll to the top on every route change. Modified 'useIntersectionObserver' in utils.jsx to accept delay as an argument, and implemented a timer to delay the intersection observer. In main.jsx, scrollRestoration setting in window's history has been changed to 'manual' to ensure the page starts from the top on a new load.